bugfix: CID-274794: Unintended sign extension
memcpy(C, sym, rq->L * rq->T);

rq->L and rq->T are both uint16_t, and this is being passed into a
size_t argument to memcpy, but this was compiling to:

movzwl  170(%rbx), %edx
movzwl  146(%rbx), %eax
imull   %eax, %edx
movslq  %edx, %rdx

Adding a cast to one of the uint16_t removes the movslq instruction:

movzwl  170(%rbx), %edx
movzwl  146(%rbx), %eax
imull   %eax, %edx

Thanks to Claudio Calvelli for helping me figure this out.

+++ src/lcrq.c
@@ -221,7 +221,7 @@ static int rq_encode_block_rfc(rq_t *rq, uint8_t *C, uint8_t *src)
 	sym = rq_decode_C(rq, &D);
 	matrix_t Cm;
 	matrix_new(&Cm, rq->L, rq->T, sym, 0);
-	memcpy(C, sym, rq->L * rq->T);
+	memcpy(C, sym, (uint32_t)rq->L * rq->T); /* cast to avoid sign-extension */
 	free(sym);
 fail:
 	matrix_free(&A);
